// shared/enum/roles.enum.ts
enum ROLES_ENUM {
  ADM_ALL = "admin.all",
  ADM = "admin",
  DASHBOARD_HOME = "dashboard.home",
  ACCOUNT_BALANCE = "account.balance",
  COMPANY_CLIENTS_ALL = "company.clients.all",
  COMPANY_PAYMENTS_ALL = "company.payments.all",
  CONFIG_COMPANY_PAYOUT = "config_company_payout",
  CONFIG_COMPANY_PAYMENT = "config_company_payment",
  CLIENT_CREATE = "client.create",
  CLIENT_UPDATE = "client.update",
  CLIENT_FIND = "client.find",
  DRIVER = "driver",
  PAYMENT_CREATE = "payment.create",
  PAYMENT_CRON = "payment.cron",
  PAYMENT_UPDATE = "payment.update",
  PAYMENT_IMPORT = "payment.import",
  PAYMENT_DELETE = "payment.delete",
  PAYMENT_FIND = "payment.find",
  PAYMENT_CREATE_QRCODE = "payment.create_pix_qrcode",
  PAYER_PAYMENT_FIND = "payer.payment.find",
  PAYOUT_TRANSFER = "payout.transfer",
  PAYOUT_STATEMENT = "payout.statement",
  WEBHOOK_PIX = "webhook.pix",
  WEBHOOK_ONBOARDING = "webhook.onboarding",
  WEBHOOK_INTERNAL_TRANSFER = "webhook.internal-transfer",
  UMA_AUTHORIZATION = "uma_authorization",
  SCHEDULE_PAYMENT_CRON = "schedule_payment.cron",
  SCHEDULE_PAYOUT_CRON = "schedule_payout.cron",
  STATEMENT_FIND = "statement.find",
  DRIVER_FIND = "driver.find",
  DRIVER_CREATE = "driver.create",
  SCHEDULE_FIND = "schedule_payment.find",
  SCHEDULE_CREATE = "schedule_payment.create",
  LOGS = "logs",
  USER_FIND = "user.find",
  USER_CREATE = "user.create",
}

export default ROLES_ENUM;